chore: build fix (circular dependency?)
