+++ src/new-client/src/components/DraggableTearOut/DraggableTearOut.tsx
@@ -110,11 +110,3 @@ const createDragImage = (
     })
   }
 }
-
-export const nonDraggableChildProps = {
-  draggable: true,
-  onDragStart: (event: React.DragEvent<HTMLDivElement>) => {
-    event.preventDefault()
-    event.stopPropagation()
-  },
-}
