+++ src/new-client/src/components/DraggableTearOut/nonDraggableChildProps.ts
@@ -1,0 +1,7 @@
+export const nonDraggableChildProps = {
+  draggable: true,
+  onDragStart: (event: React.DragEvent<HTMLDivElement>) => {
+    event.preventDefault()
+    event.stopPropagation()
+  },
+}
